+++ desmume/src/SPU.h
@@ -58,8 +58,8 @@ typedef struct
    u32 length;
    s8 *buf8;
    s16 *buf16;
-   double sampcnt;
-   double sampinc;
+   float sampcnt;
+   float sampinc;
    // ADPCM specific
    int lastsampcnt;
    s16 pcm16b;
